rating added

+++ app/src/main/java/com/example/wheretoeat/Restaurant.java
@@ -4,7 +4,7 @@ public class Restaurant {
     public String restName;
     public String restLocation;
     public String restPrice;
-    public int restRating;
+    public double restRating;
     public String restCuisineType;
     public String restWebsite;
     public String restPhoneNumber;
@@ -13,7 +13,7 @@ public class Restaurant {
 
     }
 
-    public Restaurant(String restName, String restLocation, String restPrice, int restRating, String restCuisineType, String restWebsite, String restPhoneNumber) {
+    public Restaurant(String restName, String restLocation, String restPrice, double restRating, String restCuisineType, String restWebsite, String restPhoneNumber) {
         this.restName = restName;
         this.restLocation = restLocation;
         this.restPrice = restPrice;
